% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to punts}
\usage{
add_metadata(punts)
}
\arguments{
\item{punts}{A tibble containing the punts to which metadata will be added}
}
\value{
A tibble \code{punts} with metadata added
}
\description{
This function adds game metadata to your play-by-play dataframe (should work for dataframes containing non-punting data too, fwiw)
}
\examples{
punts_with_metadata <- punts \%>\% add_metadata()
}
