% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini2.R
\name{by_punters}
\alias{by_punters}
\title{Summarize data for player careers}
\usage{
by_punters(punts, ..., threshold = 64)
}
\arguments{
\item{punts}{The play-by-play punting data to be summarized}

\item{...}{Any additional arguments will be passed through to \code{dplyr::summarise}}

\item{threshold}{The minimum number of career punts needed to be included, defaults to 64}
}
\value{
A tibble \code{punters} where each row is a punter and each column is a stat
}
\description{
This function is essentially a convenient wrapper for \code{dplyr::summarise}
which provides a dataframe with the following columns: NumPunts, pEPA, Gross, Net, RERUN, SHARP_RERUN_OF, SHARP_RERUN_PD,
first_year, last_year, team_logo_espn, team_color, team_color2. Additional columns may be added via valid calls to \code{dplyr::summarise}
as additional arguments.
For comparison of punter seasons, see \code{puntr::by_punter_seasons}; for comparison of punter games, see \code{puntr::by_punter_games}.
}
\examples{
\dontrun{
create_mini(punts)
}
}
