% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_all}
\alias{calculate_all}
\title{Calculate metrics}
\usage{
calculate_all(punts)
}
\arguments{
\item{punts}{The already-processed play-by-play punting data to which metrics will be added}
}
\value{
A tibble \code{punts} with the calculated metrics added
}
\description{
Calculate metrics
}
\examples{
\dontrun{
calculate_all(punts)
}
}
