% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college.R
\name{college_to_pro}
\alias{college_to_pro}
\title{Convert college data to \code{puntr} format}
\usage{
college_to_pro(punts, power_five = TRUE)
}
\arguments{
\item{punts}{A data frame containing punts in the cfbfastR format}

\item{power_five}{Logical, defaults to TRUE to include only punters from Power 5 teams}
}
\value{
A tibble \code{punts} in a format usable for \code{puntr::calculate_all}
}
\description{
Rename columns and process data such that the output can be plugged directly into \code{puntr::calculate_all},
and the output of that can be plugged directly into \code{puntr::create_mini} (or \code{puntr::create_miniY}).
}
\examples{
\dontrun{
college_to_pro(punts)
}
}
