% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_metadata}
\alias{import_metadata}
\title{Import metadata}
\usage{
import_metadata()
}
\value{
A tibble \code{metadata} containing metadata. Each row is a game.
}
\description{
This work piggybacks on that of greerre on github, see here: (https://github.com/Puntalytics/pfr_metadata_pull)
The pfr_metadata_pull package contains code for scraping the metadata yourself, if you're so inclined
}
\examples{
metadata <- import_metadata()
}
