% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{punt_trim}
\alias{punt_trim}
\title{Filter columns relevant to punting}
\usage{
punt_trim(punts, columns = NULL)
}
\arguments{
\item{punts}{The play-by-play punting data}

\item{columns}{Defaults to \code{NULL} indicating trimming down to the default columns. If \code{columns} is a list of other columns, those will be additionally included}
}
\value{
A tibble \code{punts} containing only the specified columns
}
\description{
Filter columns relevant to punting
}
\examples{
\dontrun{
punt_trim(punts)
punt_trim(punts, columns=c("solo_tackle", "assist_tackle"))
}
}
