% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{trust_the_process}
\alias{trust_the_process}
\title{Clean and process play-by-play punting data}
\usage{
trust_the_process(punts, trim = TRUE)
}
\arguments{
\item{punts}{The play-by-play punting data to be cleaned and processed}

\item{trim}{Specify \code{trim=FALSE} if you would like to use puntr::punt_trim to create custom columns}
}
\value{
A tibble \code{punts} with cleaned and processed play-by-play punting data
}
\description{
Clean and process play-by-play punting data
}
\examples{
punts <- trust_the_process(punts)
punts <- trust_the_process(punts_custom_trimmed_already, trim=FALSE)
}
