% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_all}
\alias{calculate_all}
\title{Calculate metrics}
\usage{
calculate_all(punts)
}
\arguments{
\item{punts}{The already-processed play-by-play punting data to which metrics will be added. Must contain at least three seasons and 1000 punts.}
}
\value{
A tibble \code{punts} with the calculated metrics added
}
\description{
This function adds all Puntalytics metrics to your punting data: RERUN, SHARP, SHARPnet and SHARP_RERUN
(plus open-field and pin-deep versions of the 3 SHARPs) as well as average and cumulative EPA (our custom EPA metric.)
This function will continue to be updated as new metrics are added.
Currently, functions for calculating individual metrics are for internal use only, and not exported.
}
\examples{
\dontrun{
calculate_all(punts)
}
}
