% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{create_mini}
\alias{create_mini}
\title{Summarize data for players}
\usage{
create_mini(punts, threshold = 64)
}
\arguments{
\item{punts}{The play-by-play punting data to be summarized}

\item{threshold}{The minimum number of career punts needed to be included, defaults to 64}
}
\value{
A tibble \code{mini} where each row is a punter and each column is a stat
}
\description{
Summarize data for players
}
\examples{
\donttest
create_mini(punts)
}
