% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{create_mini}
\alias{create_mini}
\title{Summarize data for players}
\usage{
create_mini(punts, ..., threshold = 64)
}
\arguments{
\item{punts}{The play-by-play punting data to be summarized}

\item{...}{Any additional arguments will be passed through to \code{dplyr::summarise}}

\item{threshold}{The minimum number of career punts needed to be included, defaults to 64}
}
\value{
A tibble \code{mini} where each row is a punter and each column is a stat
}
\description{
This function is essentially a convenient wrapper for \code{dplyr::summarise}
which includes all of the relevant columns from a \code{puntr}-style data frame.
NOTE: \code{puntr::create_mini}, \code{puntr::create_miniY}, and \code{puntr::create_miniG} are being phased out
in favor of \code{puntr::by_punters}, \code{puntr::by_punter_seasons}, and \code{puntr::by_punter_games}
}
\examples{
\dontrun{
create_mini(punts)
}
}
