% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{create_miniY}
\alias{create_miniY}
\title{Summarize data for player-seasons}
\usage{
create_miniY(punts, threshold = 32)
}
\arguments{
\item{punts}{The play-by-play punting data to be summarized}

\item{threshold}{The minimum number of punts for a season to be included, defaults to 32}
}
\value{
A tibble \code{miniY} where each row is a punter-season and each column is a stat
}
\description{
Summarize data for player-seasons
}
\examples{
\donttest
create_miniY(punts)
}
