% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college.R
\name{import_college_punts}
\alias{import_college_punts}
\title{Import college punting data}
\usage{
import_college_punts(years)
}
\arguments{
\item{years}{A year or range of years to be scraped}
}
\value{
A tibble \code{punts} of punts in the \code{cfbscrapR} format
}
\description{
Import college punting data
}
\examples{
\dontrun {
import_college_punts(2018:2019)
}
}
