% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college.R
\name{import_college_punts}
\alias{import_college_punts}
\title{Import college punting data}
\usage{
import_college_punts(years, weeks = 1:15)
}
\arguments{
\item{years}{A year or range of years to be scraped}

\item{weeks}{A week or range or weeks, defaults to \code{1:15}. Use fewer weeks to scrape an in-progress season}
}
\value{
A tibble \code{punts} of punts in the \code{cfbscrapR} format
}
\description{
Import college punting data for seasons in the scope of \code{cfbscrapR}.
The default \code{weeks} parameter will work for completed seasons; ask for fewer weeks to scrape an in-progress season.
Scraping both completed and in-progress seasons will require two separate function calls (see examples). This function is a
wrapper around \code{cfbscrapR::cfb_pbp_data}
}
\examples{
\dontrun{
import_college_punts(2018:2019) # completed seasons
import_college_punts(2020, weeks = 1) # just week 1 of the 2020 season
}
}
