% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_metadata}
\alias{import_metadata}
\title{Import metadata}
\usage{
import_metadata()
}
\value{
A tibble \code{metadata} containing metadata
}
\description{
Import metadata
}
\examples{
metadata <- import_metadata()
}
