% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_punts}
\alias{import_punts}
\title{Import punting data}
\usage{
import_punts(years)
}
\arguments{
\item{years}{A year or range of years between 1999 and 2019, inclusive}
}
\value{
A tibble \code{punts} containing play-by-play punting data for the specified years
}
\description{
Import punting data
}
\examples{
\donttest
import_punts(1999:2019)
}
